added minimax player version 8 which uses floats instead of doubles (should be faster with division)

+++ src/ConnectFour/BusinessLogic/BoardHeuristic.java
@@ -3,7 +3,8 @@ package src.ConnectFour.BusinessLogic;
 public interface BoardHeuristic {
     /**
      * Evaluates the given board for the given player
+     *
      * @return Returns a double between 1 and 0 representing the winning chances of the given player
      */
-    double evaluateBoard();
+    float evaluateBoard();
 }
